+++ tinyvv/query.py
@@ -3,32 +3,51 @@ import os.path as osp
 
 
 def lake_schema(LAKE):
-    return pl.scan_parquet(f'{LAKE}/annotations/*.parquet').collect_schema()
+    ann = pl.scan_parquet(f'{LAKE}/annotations/*.parquet')
+    # Rename cols with '.' inside, cuz not supported:
+    rename_dict = {c:c.replace('.', '_') for c in ann.collect_schema().names() if '.' in c}
+    return ann.rename(rename_dict).collect_schema()
 
 
 def lake_data(LAKE, samples_list, cols_list=None):
-    list_in = [f"{LAKE}/genotypes/samples/{s}.parquet" for s in samples_list]
-    sample_parquets = { f"{osp.splitext(osp.basename(x))[0]}":pl.scan_parquet(x) for x in list_in }
+    # Build a list of 'genotypes' pq, with renamed cols for join:
+    pqs_list = []
+    for s in samples_list:
+        pq_path = f"{LAKE}/genotypes/samples/{s}.parquet"
+        selected_cols = ['id','gt', 'ad']
+        rename_dict = { c:f'format_{s}_{c.upper()}' for c in selected_cols if c != 'id' }
+        pq_to_join = pl.scan_parquet(pq_path).select(selected_cols).rename(rename_dict)
+        pqs_list.append(pq_to_join)
 
-    # Same but add 'variants' and 'ann' for context passing:
-    all_parquets = { f"{osp.splitext(osp.basename(x))[0]}":pl.scan_parquet(x) for x in list_in }
+    # Full join on id:
+    # ENH: Find a way to do that in SQL bellow ???
+    joint_gt = pqs_list[0]  # Init to 1st pq
+    for pq in pqs_list[1:]:
+        joint_gt = joint_gt.join(
+            pq,
+            how='full',
+            on='id',
+            coalesce=True,
+            maintain_order='left_right'
+        )
+    all_parquets = { 'joint_gt': joint_gt }
+
+    # Add 'variants' for context passing:
     all_parquets["v"] = pl.scan_parquet(f'{LAKE}/uniq_variants/*.parquet')
-    all_parquets["ann"] = pl.scan_parquet(f'{LAKE}/annotations/*.parquet')
+    # Same for annot but 1st rename cols with '.' inside, cuz not supported:
+    ann = pl.scan_parquet(f'{LAKE}/annotations/*.parquet')
+    rename_dict = {c:c.replace('.', '_') for c in ann.collect_schema().names() if '.' in c}
+    all_parquets["ann"] = ann.rename(rename_dict)
 
     # Register all lf in global namespace: ctx = pl.SQLContext(register_globals=True)
     ctx = pl.SQLContext(frames=all_parquets)
     
-    first_sample = list(sample_parquets.keys())[0]
-    gt_cols = ','.join([f"{x}.gt AS format_{x}_GT" for x in sample_parquets.keys()])
+    gt_cols = ','.join([f"format_{x}_GT" for x in samples_list])
     if cols_list:
         ann_cols = ','.join([a for a in cols_list])
     else:
         ann_cols = ','.join([a for a in all_parquets["ann"].collect_schema().names() if not a.endswith('id')])
-    join_other_samples = '\n'.join([ f"LEFT JOIN {other} ON {first_sample}.id={other}.id" for other in list(sample_parquets.keys())[1:] ])
 
-    # MEMOs:
-    # * Diff colnames between variantplan and vcf2pq (colname vs info_colname)
-    # * Cols like "info_Gene.refGene" raise and error (caused by '.')
     query_lf = f"""
     SELECT
         chr as chromosome,
@@ -37,12 +56,28 @@ def lake_data(LAKE, samples_list, cols_list=None):
         alt AS alternate,
         {gt_cols},
         {ann_cols},
-    FROM {first_sample}
-    {join_other_samples}
+
+        FROM joint_gt
             LEFT JOIN v
-            ON {first_sample}.id=v.id
+            ON id=v.id
             LEFT JOIN ann
-            ON {first_sample}.id=ann.id
+            ON id=ann.id
     """
     print(query_lf)
     return ctx.execute(query_lf)
+
+
+def count_occurr(LAKE):
+    # Compute occurrences of all variants from all samples in lake
+    # MEMO: Keep using 'genotypes/*' instead of 'variants/*'
+    #       So that can have list of samples supporting occurrence
+    ctx = pl.SQLContext(frames={'all_samples': pl.scan_parquet(f"{LAKE}/genotypes/samples/*.parquet")})
+    query_occurr = """
+    SELECT
+        id,
+        count(*),
+        STRING_AGG(sample),
+    FROM all_samples
+    GROUP BY id;
+    """
+    return ctx.execute(query_occurr)